requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
requests==2.32.3
pypdf==5.1.0
pdfplumber==0.11.5
python-multipart==0.0.20
supabase==2.10.0
tiktoken==0.9.0
python-docx==1.1.2
PyJWT==2.10.1
argon2-cffi==23.1.0
boto3==1.35.99
pytest==8.3.4
ruff==0.8.6
email-validator==2.2.0
